requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
scipy
google-generativeai